Crud categoriaTextil, listados, fraccionALADI

+++ vucem-catalogos-service/src/main/java/com/example/vucem_catalogos_service/persistence/repo/ICatTipoTramiteRepository.java
@@ -4,6 +4,6 @@ import com.example.vucem_catalogos_service.model.entity.CatTipoTramite;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ICatTipoTramiteRepository extends JpaRepository<CatTipoTramite, Integer>,
+public interface ICatTipoTramiteRepository extends JpaRepository<CatTipoTramite, Long>,
         JpaSpecificationExecutor<CatTipoTramite> {
 }
